aws/app/requirements.txt
boto3
fastapi
mangum
llama-cpp-python
